requirements.txt
streamlit
google-genai
sentence-transformers
torch
numpy
